+++ src/gc.rs
@@ -160,6 +160,7 @@ pub(crate) fn trace(value: &Value, references: &mut Vec<usize>) {
     while let Some(value) = pending.pop() {
         match value {
             Value::Delegate(d) => pending.extend(d.receiver.as_deref()),
+            Value::ObjectReference(object) => references.push(object.allocation_id()),
             Value::SlotReference(reference)
             | Value::SlotInterface {
                 receiver: reference,